--- .env.example
@@ -1,1 +1,0 @@
-BURGER_API_URL=https://norma.nomoreparties.space/api
