+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "react-canonical",
+  "author": "Armen Zakharyan",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -92,7 +93,9 @@
     "build-storybook": "storybook build",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx ./src",
     "lint:fix": "npm run lint -- --fix",
-    "format": "prettier ./src --write"
+    "format": "prettier ./src --write",
+    "test": "jest",
+    "cypress:open": "cypress open"
   },
   "eslintConfig": {
     "extends": [
