+++ src/pages/forgot-password/forgot-password.tsx
@@ -1,7 +1,6 @@
 import { FC, useState, SyntheticEvent, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { forgotPasswordApi } from '@api';
 import { ForgotPasswordUI } from '@ui-pages';
 
 import { useSelector, useDispatch } from '@store';
