+++ src/services/store.ts
@@ -1,10 +1,10 @@
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 
-import userReducer from './slices/user-slice';
-import feedReducer from './slices/feed-slice';
-import ingredientsReducer from './slices/ingredients-slice';
-import constructorReducer from './slices/constructor-slice';
-import orderReducer from './slices/order-slice';
+import userReducer from '../slices/user-slice';
+import feedReducer from '../slices/feed-slice';
+import ingredientsReducer from '../slices/ingredients-slice';
+import constructorReducer from '../slices/constructor-slice';
+import orderReducer from '../slices/order-slice';
 
 import {
   TypedUseSelectorHook,
