+++ tsconfig.json
@@ -24,7 +24,7 @@
       "@ui-pages": ["src/components/ui/pages"],
       "@utils-types": ["src/utils/types"],
       "@api": ["src/utils/burger-api.ts"],
-      "@slices": ["src/services/slices"],
+      "@slices": ["src/slices"],
       "@hooks": ["src/hooks"],
       "@store": ["src/services/store.ts"],
       "@selectors": ["src/services/selectors"]
